User guide pre-review cleanup and code quality improvements
- Split content.tsx (2579 lines) into per-section files under content/
- Extract PageSection and PageContent into content/types.ts
- Replace React.createElement heading pattern with typed HeadingTag variable
- Move inline styles to CSS classes (ug-image, ug-icon-sm, ug-icon-md, ug-feature-grid--wide)
- Add Escape key handler to mobile menu for accessibility
- Add tabIndex and aria-label to scrollable tables for keyboard accessibility
- Add TODO comments in place of commented-out JSX blocks
- Remove leftover empty intro: '' fields now that field is optional
- Revert global.css h2/h3 changes; scope heading sizes to User Guide via DocPage.module.css
- Document ug-* global class convention in DocPage.module.css
- Document headings field intent in nav.ts
- All 562 core and 52 web tests passing
