+++ packages/core/styles/global.css
@@ -111,15 +111,15 @@ h1 {
 }
 
 h2 {
-    font-size: 28px;
+    font-size: 26px;
     font-weight: 300;
     margin: 0;
     padding: 0;
 }
 
 h3 {
-    font-size: 20px;
-    font-weight: 400;
+    font-size: 18px;
+    font-weight: 500;
     margin: 0;
     padding: 0;
 }
